v2
Partial rewrite of the program, which offers much more customisation.

- A list of messages can be added which will be selected from at random
is now possible to accompany the image within the Tweet. This supports
the new 280 character limit.

- An optional filepath can be specified for the image location.

- Saving is now optional, if disabled, images will be deleted from the
local storage after posting online is successful

- Twitter credentials and other program settings can now be entered into
the setup.py file. Allowing the main program script to simply import the
settings and not require the user to enter this file themselves.

Read through the setup.py or the readme to see an explanation of all the
newly added settings